Remove play-json and move to https://github.com/playframework/play-json
