--- framework/src/play-json/src/main/scala/play/api/libs/json/Util.scala
@@ -1,16 +1,0 @@
-/*
- * Copyright (C) 2009-2016 Lightbend Inc. <https://www.lightbend.com>
- */
-package play.api.libs.json.util
-
-import scala.language.higherKinds
-
-trait LazyHelper[M[_], T] {
-  def lazyStuff: M[T]
-}
-
-object LazyHelper {
-  def apply[M[_], T](stuff: M[T]) = new LazyHelper[M, T] {
-    override lazy val lazyStuff = stuff
-  }
-}
